Fix case-sensitive file issue
