+++ src/main/java/umc/hackathon/service/DiaryService/DiaryCommandService.java
@@ -9,5 +9,5 @@ public interface DiaryCommandService {
 
     Diary updateDiary(Long userId, Long diaryId, DiaryRequestDTO.UpdateDiaryDTO request);
 
-    //void deleteDiary(Long userId, Long diaryId);
+    void deleteDiary(Long userId, Long diaryId);
 }
