+++ src/main/java/umc/hackathon/service/MyPlantService/MyPlantService.java
@@ -12,4 +12,7 @@ public interface MyPlantService {
 
     // 반려식물 정보조회(== 메인화면)
     MyPlantResponseDTO.MyPlantDTO getMyPlant(Long myPlantId);
+
+    // 반려식물 삭제
+    void deleteMyPlant(Long myPlantId);
 }
