+++ src/main/java/umc/hackathon/web/dto/MyPlant/MyPlantResponseDTO.java
@@ -1,15 +1,15 @@
-package umc.hackathon.web.dto.Myplant;
+package umc.hackathon.web.dto.MyPlant;
 
 import lombok.*;
 
-public class MyplantResponseDTO {
+public class MyPlantResponseDTO {
 
     @Builder
     @Getter
     @Setter
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class MyplantDTO {
+    public static class MyPlantDTO {
         private String nickname;
         private String imageUrl;
         private Integer status;
