--- src/main/java/umc/hackathon/web/dto/Myplant/MyplantResponseDTO.java
@@ -1,18 +1,0 @@
-package umc.hackathon.web.dto.MyPlant;
-
-import lombok.*;
-
-public class MyPlantResponseDTO {
-
-    @Builder
-    @Getter
-    @Setter
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class MyPlantDTO {
-        private String nickname;
-        private String imageUrl;
-        private Integer status;
-
-    }
-}
